# server/docs/swagger.yaml
basePath: /
definitions:
  api.GetChatRequest:
    properties:
      from:
        type: integer
      page:
        default: 1
        type: integer
      pageSize:
        default: 10
        type: integer
      to:
        type: integer
    type: object
...
info:
  contact: {}
  title: MyTodo
  version: "1.0"
paths:
...
  /chat/del:
    post:
      description: delete chat
      parameters:
      - description: ID
        in: formData
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Del Chat
      tags:
      - Chat
...
  /chat/get:
    post:
      description: get chat
      parameters:
      - description: GetChatRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.GetChatRequest'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get Chat
      tags:
      - Chat
...
swagger: "2.0"